+++ src/main/java/org/zalando/nakadi/enrichment/Enrichment.java
@@ -8,7 +8,7 @@ import org.zalando.nakadi.domain.EventCategory;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.EventTypeBase;
 import org.zalando.nakadi.exceptions.EnrichmentException;
-import org.zalando.nakadi.exceptions.InvalidEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.InvalidEventTypeException;
 
 import java.util.HashSet;
 import java.util.Set;
