--- src/main/java/org/zalando/nakadi/exceptions/InvalidEventTypeException.java
@@ -1,17 +1,0 @@
-package org.zalando.nakadi.exceptions;
-
-import org.zalando.problem.MoreStatus;
-
-import javax.ws.rs.core.Response;
-
-public class InvalidEventTypeException extends NakadiException {
-
-    public InvalidEventTypeException(final String message) {
-        super(message);
-    }
-
-    @Override
-    protected Response.StatusType getStatus() {
-        return MoreStatus.UNPROCESSABLE_ENTITY;
-    }
-}
